requirements.txt
torch
torchvision
numpy
Pillow
PyYAML
tqdm
matplotlib
opencv-python
transformers
pypdf
